sync-github-releases: extract changelog parsing into changelog.ts
parseChangelog (and its ChangelogSections type) is a self-contained concern —
turning a release-me CHANGELOG.md into per-version notes — that has nothing to
do with orchestration. Move it out of the index.ts grab-bag into its own file.
